Make DateTime opaque
I'm planning a replacement to chrono, pending they don't fix a major
vulnerability with Utc.now() calls (race condition), due to depending on
an old version of the `time` crate (0.1).

`time` 0.3 or above are supposed to have fixed this problem. I'm
evaluating the situation. Meanwhile, I decided to prepare for a possible
migration.